+++ common/src/main/java/io/github/mortuusars/scholar/client/gui/screen/edit/LecternSpreadBookEditScreen.java
@@ -20,6 +20,7 @@ import net.minecraft.world.item.ItemStack;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
+import java.util.ArrayList;
 import java.util.Optional;
 
 public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements MenuAccess<LecternSpreadBookEditMenu> {
@@ -115,7 +116,8 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
     @Override
     protected void sendChanges(@Nullable String title) {
         removeEmptyTrailingPages();
-        Packets.sendToServer(new LecternEditBookC2SP(getMenu().getLecternPos(), pages, Optional.ofNullable(title)));
+        // Copying 'pages' list to not cause ConcurrentModificationException when packet is encoded:
+        Packets.sendToServer(new LecternEditBookC2SP(getMenu().getLecternPos(), new ArrayList<>(pages), Optional.ofNullable(title)));
     }
 
     // --
@@ -188,7 +190,7 @@ public class LecternSpreadBookEditScreen extends SpreadBookEditScreen implements
 
     protected void sendButtonClick(int buttonId) {
         if (Minecraft.getInstance().gameMode != null) {
-//            saveChanges(false, null);
+            saveChanges(false, null);
             Minecraft.getInstance().gameMode.handleInventoryButtonClick(this.menu.containerId, buttonId);
         }
     }