+++ common/src/main/java/io/github/mortuusars/scholar/client/gui/screen/edit/SpreadBookEditScreen.java
@@ -382,7 +382,8 @@ public class SpreadBookEditScreen extends SpreadBookScreen {
 
     protected void enterSignMode() {
         saveChanges(false, null);
-        minecraft.setScreen(new BookSigningScreen(this, bookColor, title -> saveChanges(true, title)));
+        minecraft.execute(() ->
+            minecraft.setScreen(new BookSigningScreen(this, bookColor, title -> saveChanges(true, title))));
     }
 
     // --
@@ -394,8 +395,13 @@ public class SpreadBookEditScreen extends SpreadBookScreen {
 
     protected void setPageText(Spread.Side side, String text) {
         int pageIndex = side.getPageIndexFromSpread(currentSpread);
-        if (pageIndex >= 0 && pageIndex < this.pages.size()) {
 
+        // Add empty pages until current page index
+        while (pageIndex > pages.size() - 1 && pageIndex < getPageCount()) {
+            appendEmptyPage();
+        }
+
+        if (pageIndex >= 0 && pageIndex < this.pages.size()) {
             String currentText = getPageText(side);
             getHistory().add(() -> {
                 pages.set(pageIndex, text);
@@ -431,7 +437,7 @@ public class SpreadBookEditScreen extends SpreadBookScreen {
         Change change = Change.create(() -> {
             pages.add(pageIndex, "");
             while (pages.size() >= 100) {
-                pages.remove(pages.size() - 1);
+                pages.removeLast();
             }
             setTextBoxes();
             bookModified = true;
@@ -620,9 +626,7 @@ public class SpreadBookEditScreen extends SpreadBookScreen {
             FileDialogs.loadFile(defaultDirectory, title, "Text Files (.txt)", false, "*.txt").ifPresent(filePath -> {
                 try {
                     String content = Files.readString(Path.of(filePath));
-                    Minecraft.getInstance().execute(() -> {
-                        setBookContents(content, withFormatting);
-                    });
+                    Minecraft.getInstance().execute(() -> setBookContents(content, withFormatting));
                 } catch (IOException e) {
                     Minecraft.getInstance().execute(() -> player.displayClientMessage(
                             Component.translatable("gui.scholar.import_book.failure"), false));